organize logic into a separate class
